refactor: MoviePreview 파일명 변경
