+++ README.md
@@ -5,6 +5,7 @@
 FE 레벨1 영화관 미션
 
 1. 🎬 영화 목록 조회 (인기순)
+<<<<<<< HEAD
    영화 목록의 1페이지를 불러오며 더보기 버튼을 누르면 그 다음의 영화 목록을 불러 올 수 있다.
    단, 페이지 끝에 도달한 경우에는 더보기 버튼을 화면에 출력하지 않는다.
    ⚠️ 인기순은 TMDB에서 제공하는 API의 속성 이름을 나타내는 것이므로 별도로 받은 데이터를 정렬하지 않습니다.
@@ -19,10 +20,27 @@ FE 레벨1 영화관 미션
 3. ⚠️ 오류
    오류가 발생하는 경우에는 사용자를 위한 오류 메시지를 띄워 준다.
    어떤 오류를 대응해야 하고, 어떤 UI로 보여줄 것인지는 자율적으로 결정한다.
+=======
+영화 목록의 1페이지를 불러오며 더보기 버튼을 누르면 그 다음의 영화 목록을 불러 올 수 있다.
+단, 페이지 끝에 도달한 경우에는 더보기 버튼을 화면에 출력하지 않는다.
+⚠️ 인기순은 TMDB에서 제공하는 API의 속성 이름을 나타내는 것이므로 별도로 받은 데이터를 정렬하지 않습니다.
+figma 시안과는 달리 20개씩 영화 목록을 보여주면 됩니다.
+영화 목록 아이템에 대한 Skeleton UI를 구현한다.
+Skeleton UI는 템플릿으로 제공되는 파일 이외로 자유롭게 구현할 수 있다.
+2. 🔎 검색
+영화 검색 API를 이용하여 내가 보고 싶은 영화를 검색할 수 있다.
+엔터키를 눌러 검색할 수 있다
+검색 버튼을 클릭하여 검색할 수 있다
+영화 목록 조회와 같이 검색한 결과에 한해 정보를 보여주는 화면의 요구사항은 동일하다
+3. ⚠️ 오류
+오류가 발생하는 경우에는 사용자를 위한 오류 메시지를 띄워 준다.
+어떤 오류를 대응해야 하고, 어떤 UI로 보여줄 것인지는 자율적으로 결정한다.
+>>>>>>> 52f8a27 (docs(README): 리드미 작성)
 
 ## UI 기능
 
 ### 컴포넌트
+<<<<<<< HEAD
 
 - [x] Header: 아이콘 및 검색창을 가지는 헤더
 - [x] SearchBar: 검색 컴포넌트
@@ -76,3 +94,19 @@ FE 레벨1 영화관 미션
 - [x] 도메인 객체 생성
 - [ ] MovieListController 쪼개기
 - [ ] 함수형 컴포넌트 클래스화 하기
+=======
+- [ ] Header: 아이콘 및 검색창을 가지는 헤더
+- [ ] SearchBar: 검색 컴포넌트
+- [ ] MoviePoster: 영화 포스터 컴포넌트 (영화 이미지, 제목, 별점을 보여준다)
+- [ ] MoviePosterBoard: 상태에 따라 영화 포스터들을 보여주는 컴포넌트
+- [ ] NetworkErrorFallback: 네트워크 에러 발생 시 보여지는 컴포넌트
+
+## UX 
+- [ ] 네트워크 요청 중: 스켈레톤 UI 보여주기
+- [ ] 네트워크 요청 실패 
+    - [ ] 최초 접속 / 검색: MoviePosterBoard내에 NetworkErrorFallback 컴포넌트를 보여준다
+    - [ ] '더보기' 버튼 클릭 시: 20개의 서스펜스가 보이다가 사라지고 alert으로 에러메시지 보여줌
+- [ ] 없는 영화를 검색했을 시: '🎬원하는 영화를 찾을 수 없습니다' 텍스트를 보여준다.
+- [ ] '더보기' 버튼 클릭 시 아래로 스크롤한다
+- [ ] 이미지 로드가 실패하면 디폴트 이미지를 보여준다
+>>>>>>> 52f8a27 (docs(README): 리드미 작성)
