+++ src/components/MoviePreview/createMoviePreview.ts
