+++ src/components/MoviePreview/createMovieThumbnail.ts
