+++ src/components/MoviePreview/style.css
