+++ src/components/MoviePreview/utils/createMovieSkeleton.ts
