+++ src/components/MoviePreview/utils/createSkeletonPreview.ts
