--- src/components/moviePoster/style.css
@@ -1,47 +1,0 @@
-.item-card {
-  display       : flex;
-  flex-direction: column;
-}
-
-.item-thumbnail {
-  border-radius  : 8px;
-  width          : 180px;
-  height         : 270px;
-  background-size: contain;
-}
-
-.item-title {
-  margin-top : 16px;
-  font-size  : 1.2rem;
-  font-weight: bold;
-}
-
-.item-score {
-  margin-top: 16px;
-  font-size : 1.2rem;
-}
-
-.item-score::after {
-  margin-left: 8px;
-}
-
-.item-title.skeleton::after,
-.item-score.skeleton::after {
-  font-size: 0;
-  content  : "loading";
-}
-
-.item-card .skeleton {
-  background     : linear-gradient(-90deg, #aaa, #f0f0f0, #aaa, #f0f0f0);
-  background-size: 400%;
-  animation      : skeleton-animation 5s infinite ease-out;
-  border-radius  : 8px;
-}
-
-.item-list-skeleton {
-  margin-top: 0
-}
-
-.item-view>.display-none {
-  display: none;
-}
